Move user analysis code in its own module
